// src/models/interface.ts
export type Id = mongoose.Types.ObjectId;
// ...
export type RoleCamp = Mode | "peto";
export type Role = RoleCamp | "admin";
export type Mode = "nong" | "pee";
// ...
export interface InterTimeOffset {
  day: number;
  hour: number;
  minute: number;
  _id: Id;
  //private
}
// ...
export interface InterChat {
  message: string;
  userId: Id;
  campModelId: Id;
  role: RoleCamp;
  typeChat: TypeChat;
  refId: Id; //'น้องคุยส่วนตัวกับพี่','คุยกันในบ้าน' baan,'คุยกันในฝ่าย' part,'พี่คุยกันในบ้าน' baan,'พี่บ้านคุยกัน' part
  campMemberCardIds: Id[];
  date: Date;
  _id: Id;
  //public
}
export interface ShowChat extends InterChat {
  nickname: string;
  baanName: string;
  partName: string;
  roomName: string;
  canReadInModeNong: boolean;
  //public
}
export interface CreatePeeChat {
  message: string;
  partId: Id;
  //public
}
export interface EditChat {
  message: string;
  id: Id;
  //public
}
export interface CreateBaanChat {
  message: string;
  baanId: Id;
  //public
}
export interface CreateNongChat {
  message: string;
  CampMemberCard: Id;
  //public
}
// ...
export const typeChats = [
  "น้องคุยส่วนตัวกับพี่",
  "คุยกันในบ้าน",
  "คุยกันในฝ่าย",
  "พี่คุยกันในบ้าน",
  "พี่บ้านคุยกัน",
] as const;
export type TypeChat = (typeof typeChats)[number];
export type GetChat =
  | "getAllChatFromCampId"
  | "getPartChat"
  | "getNongBaanChat"
  | "getPeeBaanChat"
  | "getNongChat"
  | "getPartPeebaanChat";
export interface AllPlaceData {
  allPlace: Map<string, InterPlace[]>;
  allBuildings: Map<Id, string>;
  //public
}
export interface ChatReady {
  chats: ShowChat[];
  timeOffset: InterTimeOffset;
  mode: Mode;
  sendType: {
    id: Id;
    roomType: TypeChat;
  } | null;
  success: boolean;
  roomName: string;
  userId: Id;
  subscribe: string;
  camp: BasicCamp;
  //private
}
// ...
export interface BasicCamp {
  nameId: Id;
  round: number;
  dateStart: Date;
  dateEnd: Date;
  nongDataLock: boolean;
  open: boolean;
  songIds: Id[];
  link: string | null;
  allDone: boolean;
  lockChangePickup: boolean;
  pictureUrls: string[];
  registerModel: "noPaid" | "noInterview" | "all";
  memberStructure:
    | "nong->highSchool,pee->1year,peto->2upYear"
    | "nong->highSchool,pee->2upYear"
    | "nong->1year,pee->2upYear"
    | "nong->highSchool,pee->allYear"
    | "allYearMix";
  logoUrl: string | null;
  registerSheetLink: string | null;
  peeLock: boolean;
  _id: Id;
  campName: string;
  nongSleepModel:
    | "นอนทุกคน"
    | "เลือกได้ว่าจะค้างคืนหรือไม่"
    | "ไม่มีการค้างคืน";
  peeSleepModel: "นอนทุกคน" | "เลือกได้ว่าจะค้างคืนหรือไม่" | "ไม่มีการค้างคืน";
  groupName: string;
  peeDataLock: boolean;
  petoDataLock: boolean;
  haveCloth: boolean;
  actionPlanOffset: number;
  mdTime: Date;
  showCorrectAnswerAndScore: boolean;
  canAnswerTheQuestion: boolean;
  mealIds: Id[];
  canNongSeeAllAnswer: boolean;
  canNongSeeAllActionPlan: boolean;
  canNongSeeAllTrackingSheet: boolean;
  canNongAccessDataWithRoleNong: boolean;
  lockChangeQuestion: boolean;
  canReadTimeOnMirror: boolean;
  nongCall: string;
  boyZoneLadyZoneState: BoyZoneLadyZoneState;
  canNongSeeBaanOrder: boolean;
  nongIds: Id[];
  peeIds: Id[];
  petoIds: Id[];
  boardIds: Id[];
  baanIds: Id[];
  partIds: Id[];
  //public
}
export interface BasicPart {
  nameId: Id;
  campId: Id;
  peeIds: Id[];
  petoIds: Id[];
  _id: Id;
  partName: string;
  peeSleepIds: Id[];
  auths: AuthType[];
  //public
}
export interface SystemInfo {
  systemMode: string;
  endEmail: string;
  studentIdLength: number;
  studentIdLastTwoDigit: string;
}
// ...
export type BoyZoneLadyZoneState = (typeof boyZoneLadyZoneStates)[number];